generate list of license information for packages
Many packages define already metadata about their license (PKG_LICENSE),
but this is only included in the ipk files.

This change allows to create the information also on the build-host,
to get an overview on the used licenses.
In the full list, also all packages without this info are shown

Signed-off-by: Thomas Langer <thomas.langer@lantiq.com>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@43070 3c298f89-4303-0410-b956-a3cf2f4a3e73